+++ src/cpp/encoder/util.cpp
@@ -2,7 +2,7 @@
 
 #include "./util.hpp"
 
-float rbf_kernel(float x, float y, float std)
+double rbf_kernel(double x, double y, double std)
 {
   return exp(-pow(x - y, 2.) / (2. * pow(std, 2.)));
 }
